Split utils test into separate files
