+++ utils/logging_utils_test.go
@@ -1,0 +1,3 @@
+package utils
+
+// TODO: Define file utils tests
